+++ Models/ApiKeyAuthenticationOptions.cs
@@ -1,0 +1,8 @@
+using Microsoft.AspNetCore.Authentication;
+
+namespace Review.Api.Models;
+
+public class ApiKeyAuthenticationOptions : AuthenticationSchemeOptions
+{
+    public string DisplayMessage { get; set; }
+}