+++ Models/ApplicationProfile.cs
@@ -1,0 +1,17 @@
+using AutoMapper;
+using Review.Api.Models.DTOs;
+
+namespace Review.Api.Models;
+
+public class ApplicationProfile : Profile
+{
+    public ApplicationProfile()
+    {
+        CreateMap<Project, ProjectDto>();
+        CreateMap<Project, ProjectWithAssetsDto>();
+        CreateMap<CreateProjectDto, Project>();
+        CreateMap<Asset, AssetDto>();
+        CreateMap<User, UserDto>();
+        CreateMap<ApiKey, ApiKeyDto>();
+    }
+}