+++ Models/Asset.cs
@@ -1,12 +1,11 @@
 namespace Review.Api.Models;
 
-public class Asset
+public class Asset : BaseEntity
 {
-    public string Id { get; set; } = Guid.NewGuid().ToString();
+    
     public string FileName { get; set; }
     public string FileUrl { get; set; }
     public string FileType { get; set; }
-    public DateTime CreatedAt { get; set; }
 
     public string? ProjectId { get; set; }
     public Project? Project { get; set; }
@@ -15,5 +14,4 @@ public class Asset
     public User? User { get; set; }
     
     public ICollection<Comment> Comments { get; } = new List<Comment>(); // Collection navigation containing dependents
-
 }