+++ Program.cs
@@ -1,22 +1,45 @@
+using System.Text.Json.Serialization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Review.Api.Contexts;
+using Review.Api.Handlers;
 using Review.Api.Middleware;
 using Review.Api.Models;
+using Review.Api.Repositories;
+using Review.Api.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 
+builder.Services.AddScoped<IUserRepository, UserRepository>();
+builder.Services.AddScoped<IApiKeyRepository, ApiKeyRepository>();
+builder.Services.AddScoped<IApiKeyService, ApiKeyService>();
+builder.Services.AddScoped<IProjectRepository, ProjectRepository>();
+builder.Services.AddScoped<IProjectService, ProjectService>();
+builder.Services.AddScoped<IApiKeySecretHasher, ApiKeySecretHasher>();
+
+builder.Services.AddAutoMapper(options =>
+{
+    options.LicenseKey = builder.Configuration["AutoMapper:ApiKey"];
+    options.AddMaps(typeof(Program));
+});
+
+const string databaseName = "ReviewDatabase";
+
+builder.Services.AddDbContext<ApplicationDbContext>(options => options.UseInMemoryDatabase(databaseName));
+
 // Add services to the container.
-builder.Services.AddControllers();
+builder.Services.AddControllers().AddJsonOptions(options =>
+{
+    options.JsonSerializerOptions.ReferenceHandler = ReferenceHandler.IgnoreCycles;
+});
+
+
 // Learn more about configuring OpenAPI at https://aka.ms/aspnet/openapi
 builder.Services.AddOpenApi();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
-const string databaseName = "ReviewDatabase";
-builder.Services.AddDbContext<AppDbContext>(options => options.UseInMemoryDatabase(databaseName));
-builder.Services.AddDbContext<IdentityContext>(options => options.UseInMemoryDatabase(databaseName));
 
 builder.Services.AddIdentityApiEndpoints<User>(options =>
 {
@@ -31,19 +54,42 @@ builder.Services.AddIdentityApiEndpoints<User>(options =>
     // TODO set owasp password standards
 
     options.User.RequireUniqueEmail = true;
-}).AddEntityFrameworkStores<IdentityContext>();
+}).AddEntityFrameworkStores<ApplicationDbContext>();
 
-builder.Services.AddAuthorization();
+builder.Services.AddAuthentication().AddScheme<ApiKeyAuthenticationOptions, ApiKeyAuthenticationHandler>(ApiKeyConstants.ApiKeyName,
+    options =>
+    {
+        options.DisplayMessage = "Api key test";
+    });
+
+
+builder.Services.AddAuthorization(options =>
+{
+    options.AddPolicy("BearerOnly", policy =>
+    {
+        policy.RequireAuthenticatedUser();
+        policy.AddAuthenticationSchemes(
+            IdentityConstants.BearerScheme
+        );
+    });
+
+    options.AddPolicy("ApiKeyOrUser", policy =>
+    {
+        policy.RequireAuthenticatedUser();
+        policy.AddAuthenticationSchemes(
+            ApiKeyConstants.ApiKeyName,
+            IdentityConstants.BearerScheme
+        );
+    });
+});
 
 var app = builder.Build();
 
 using (var scope = app.Services.CreateScope())
 {
-    var context = scope.ServiceProvider.GetRequiredService<AppDbContext>();
+    var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
     context.Database.EnsureCreated();
 
-    var identityContext = scope.ServiceProvider.GetRequiredService<IdentityContext>();
-    identityContext.Database.EnsureCreated();
 
     // TODO remove before deployment
     var userManager = scope.ServiceProvider.GetRequiredService<UserManager<User>>();
@@ -55,6 +101,7 @@ using (var scope = app.Services.CreateScope())
     {
         User user = new User
         {
+            Id = "USER0-0000-0000-0000-000000000001",
             UserName = email,
             Email = email,
             EmailConfirmed = true,